Add debugging output for cache lookups
This commit adds printing (to stdout) for various cache
related operations. The cache does not work and this output
is to help find why. It appears the cache is not properly
carried around (cache entries are lost between locations),
but at least this helps identify the problem and give output.